[broken] move chainsync
