+++ simulation/ouroboros-leios-sim.cabal
@@ -1,4 +1,4 @@
-cabal-version:      >=1.10
+cabal-version:      3.0
 
 -- Initial package description 'ouroboros-leios-sim.cabal' generated by
 -- 'cabal init'.  For further documentation, see
@@ -33,6 +33,7 @@ library
     PlotTCP
     PraosProtocol
     PraosProtocol.BlockFetch
+    PraosProtocol.ChainSync
     PraosProtocol.Types
     RelayProtocol
     SimRelay
@@ -50,7 +51,7 @@ library
 
   -- other-extensions:
   build-depends:
-      array
+    , array
     , base
     , cairo
     , Chart
@@ -81,7 +82,7 @@ library
 executable ouroboros-net-vis
   main-is:          src/Main.hs
   build-depends:
-      base
+    , base
     , optparse-applicative
     , ouroboros-leios-sim
